chore: rename health file to cluster
