+++ lib/esse/transport/cluster.rb
